# src/main/resources/static/api-docs.yaml
openapi: 3.1.0
info:
  title: SocialMedia OpenApi Definition
  description: Open Api for SocialMedia project
  version: '1'
servers:
  - url: http://127.0.0.1:8081/api
    description: Generated server url
...
paths:
  /p/setting:
    put:
      tags:
        - Profile Settings
      summary: Set user profile information
      description: Set user's First name, Last name, country, Date of Birth and Accessibility. Before this operation the account is considered invalid.
      operationId: setting
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: success
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserErrorResponse'
      security:
        - bearerAuth: []
  /posts/like/{id}:
    post:
      tags:
        - Posts
      summary: Like post (id)
      description: Returns the new value of likes
      operationId: like
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: 1
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Post not found!
      security:
        - bearerAuth: []
  /posts/:
    post:
      tags:
        - Posts
      summary: Create a new post
      description: Create a new post with text and image (image may be null)
      operationId: create
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                text:
                  type: string
                image:
                  type: string
                  format: binary
              required:
                - image
                - text
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: success
        '500':
          description: INTERNAL_SERVER_ERROR
          content:
            text/plain:
              schema:
                type: string
                example: Error saving file
      security:
        - bearerAuth: []
...
  /search/{page}:
    get:
      tags:
        - Search
      summary: Find 10 most relevant users by page
      operationId: searchDocuments
      parameters:
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSearchDocument'
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: Too large request
        '429':
          description: TOO_MANY_REQUESTS
          content:
            text/plain:
              schema:
                type: string
                example: Too many request, please wait...
      security:
        - bearerAuth: []
  /search/fast:
    get:
      tags:
        - Search
      summary: Find 5 most relevant users
      description: Search users by query without pagination
      operationId: fastSearchDocuments
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSearchDocument'
        '400':
          description: BAD_REQUEST
          content:
            text/plain:
              schema:
                type: string
                example: Too large request
        '429':
          description: TOO_MANY_REQUESTS
          content:
            text/plain:
              schema:
                type: string
                example: Too many request, please wait...
      security:
        - bearerAuth: []
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Get post by id
      operationId: post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PostResponseDTO'
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Post not found!
  /posts/users/{id}:
    get:
      tags:
        - Posts
      summary: Get all posts by userId
      operationId: getAllPostsByUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponseDTO'
                uniqueItems: true
  /p/users/{id}:
    get:
      tags:
        - Profile Settings
      summary: Get user data by id
      description: Get UserDto by user Id. If user is private and not your friend, private fields will be null.
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserErrorResponse'
  /p/status:
    get:
      tags:
        - Profile Settings
      summary: Status between users
      description: Get the status between the authorized user and the specified user (id). Status could be BLACKLIST, BLACKLISTED, FRIENDS, INVITE, INVITED, UNKNOWN, BLOCKED
      operationId: getStatus
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: INVITE
        '400':
          description: BAD_REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserErrorResponse'
      security:
        - bearerAuth: []
  /p/me:
    get:
      tags:
        - Profile Settings
      summary: Get id of the Authenticated user
      operationId: me
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
                example: 1
      security:
        - bearerAuth: []
...
  /admin/posts/{id}:
    delete:
      tags:
        - Admin Functions
      summary: Delete post
      description: Delete post (id) if it has violated the rules.
      operationId: deletePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: Post 1 is deleted
        '404':
          description: NOT_FOUND
          content:
            text/plain:
              schema:
                type: string
                example: Post not found!
      security:
        - bearerAuth: []
components:
  schemas:
    UserErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Error message
    SettingRequestDTO:
      type: object
      properties:
        firstname:
          type: string
          example: TestFirstName
          minLength: 1
        lastname:
          type: string
          example: TestLastName
          minLength: 1
        country:
          type: string
          example: TestCountry
        dateOfBirth:
          type: string
          format: date-time
        isPrivate:
          type: boolean
          example: false
      required:
        - country
        - dateOfBirth
        - firstname
        - lastname
...
    UserSearchDocument:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        fullName:
          type: string
        profilePicture:
          type: string
        country:
          type: string
        private:
          type: boolean
    PostResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        photoUrl:
          type: string
        text:
          type: string
        countOfLike:
          type: integer
          format: int64
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        dateOfBirth:
          type: string
        profilePicture:
          type: string
        country:
          type: string
        private:
          type: boolean
        valid:
          type: boolean
        blocked:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT